corrige tipos de variáveis e formatação de entrada/saída no cálculo de média

+++ calculo_de_media.c
@@ -1,23 +1,23 @@
 #include <stdio.h>
 
 int main() {
-    int nota1, nota2, nota3;
-    int media;
+    float nota1, nota2, nota3;
+    float media;
 
     printf("*** Programa de Cálculo de Média ***\n");
 
     printf("Digite a sua primeira nota: \n");
-    scanf("%d", &nota1);
+    scanf("%f", &nota1);
 
     printf("Digite a sua segunda nota: \n");
-    scanf("%d", &nota2);
+    scanf("%f", &nota2);
 
     printf("Digite a sua terceira nota: \n");
-    scanf("%d", &nota3);
+    scanf("%f", &nota3);
 
     media = (nota1 + nota2 + nota3) / 3;
 
-    printf("A média é: %d", media);
+    printf("A média é: %f", media);
 
     return 0;
 }
